requirements.txt
seaborn==0.13.2
scanpy==1.10.3
numpy==1.26.3
torch==1.13.0
annoy==1.17.3
scipy>=1.9.0
pandas>=1.5.0
scikit-learn>=1.1.0
umap-learn>=0.5.0
matplotlib>=3.5.0
tqdm>=4.64.0 